django-backend/app/requirements.txt
Django==3.0.5
packaging==20.4
gunicorn==20.0.4
sqlparse==0.2.4
django-cors-headers==3.5.0
pyparsing==2.4.7
Jinja2==2.11.2
psycopg2-binary
django-rest-framework==0.1.0
drf-yasg==1.17.1
python-environ
djangorestframework==3.11.1